+++ fitness-tracker-client/src/app/app.component.ts
@@ -2,8 +2,8 @@ import { Component, OnInit } from '@angular/core';
 import { RouterOutlet } from '@angular/router';
 import { OptionsComponent } from './components/options/options.component';
 import { HeaderComponent } from './components/header/header.component';
-import { AuthService } from './services/auth.service';
-import { UserStoreService } from './services/user-store.service';
+import { AuthService } from './core/services/auth.service';
+import { UserStoreService } from './core/services/user-store.service';
 import { CommonModule } from '@angular/common';
 
 @Component({
