+++ fitness-tracker-client/src/app/app.routes.ts
@@ -2,7 +2,7 @@ import { Routes } from '@angular/router';
 import { LoginComponent } from './components/login/login.component';
 import { SignupComponent } from './components/signup/signup.component';
 import { DashboardComponent } from './components/dashboard/dashboard.component';
-import { AuthGuard } from './guards/auth.guard';
+import { AuthGuard } from './core/guards/auth.guard';
 import { UserManagementComponent } from './components/management/user-management/user-management.component';
 import { WorkoutTypesManagementComponent } from './components/management/workout-types-management/workout-types-management.component';
 import { WorkoutsManagementComponent } from './components/management/workouts-management/workouts-management.component';
