+++ fitness-tracker-client/src/app/components/dashboard/dashboard.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
-import { AuthService } from '../../services/auth.service';
+import { AuthService } from '../../core/services/auth.service';
 import { CommonModule } from '@angular/common';
-import { UserStoreService } from '../../services/user-store.service';
+import { UserStoreService } from '../../core/services/user-store.service';
 
 @Component({
   selector: 'app-dashboard',
