+++ fitness-tracker-client/src/app/components/login/login.component.ts
@@ -1,10 +1,10 @@
 import { Component, OnInit } from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { ReactiveFormsModule } from '@angular/forms';
-import { AuthService } from '../../services/auth.service';
+import { AuthService } from '../../core/services/auth.service';
 import { Router } from '@angular/router';
-import { TokenDto } from '../../models/tokenDto';
-import { UserStoreService } from '../../services/user-store.service';
+import { TokenDto } from '../../core/models/tokenDto';
+import { UserStoreService } from '../../core/services/user-store.service';
 
 @Component({
   selector: 'app-login',
