+++ fitness-tracker-client/src/app/components/options/options.component.ts
@@ -1,8 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { Router } from '@angular/router';
-import { UserStoreService } from '../../services/user-store.service';
-import { AuthService } from '../../services/auth.service';
+import { UserStoreService } from '../../core/services/user-store.service';
+import { AuthService } from '../../core/services/auth.service';
 
 @Component({
   selector: 'app-options',
