+++ fitness-tracker-client/src/app/core/guards/auth.guard.ts
