+++ fitness-tracker-client/src/app/core/services/auth.service.ts
