+++ fitness-tracker-client/src/app/core/services/user-store.service.ts
